pyproject.toml
[project]
name = "lieika-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Elisey",email = "eliseyblajen@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.86.0,<2.0.0)",
    "sqladmin (>=0.20.1,<0.21.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
